src/pipes/validation-pipe.pipe.ts
import {
  ArgumentMetadata,
  BadRequestException,
  Injectable,
  PipeTransform,
} from '@nestjs/common'
import { plainToInstance } from 'class-transformer'
import { validate } from 'class-validator'

@Injectable()
export class ValidationPipe implements PipeTransform {
  async transform(value: any, { metatype }: ArgumentMetadata) {
    if (!metatype || !this.toValidate(metatype)) {
      return value
    }

    const object = plainToInstance(metatype, value)
    const errors = await validate(object)

    if (errors.length > 0) {
      throw new BadRequestException(errors.join('\n'))
    }

    return value
  }

  private toValidate(metatype: new () => any): boolean {
    const types: Array<new () => any> = [String, Boolean, Number, Array, Object]
    return !types.includes(metatype)
  }
}
